requirements.txt
python-dotenv==1.1.1
Requests==2.32.5
rich==14.1.0
supabase==2.18.1
urllib3==2.5.0
